pyproject.toml
[project]
name = "skincancerdetection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.2",
    "uvicorn[standard]>=0.40.0",
    "tensorflow==2.15.1",
    "numpy>=1.24.0",
    "pillow>=12.1.0",
    "pydantic>=2.5.0",
    "starlette>=0.37.0",
    "python-multipart>=0.0.9",
    "typing-extensions>=4.8.0",
    "pydantic-settings>=2.2.0",
]
[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0"
]


